# OpenAPI.yml
...
openapi: 3.0.3
info:
  title: Drone Feeder
  version: 1.0.0-SNAPSHOT
paths:
  /drone:
    post:
      tags:
        - Drone
      description: Create a delivery
      operationId: addDrone
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DroneDto'
      responses:
        '201':
          description: Create a drone
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Drone'
    get:
      tags:
        - Drone
      description: Returns all registered drones
      operationId: getDrones
      responses:
        '200':
          description: Get all the drones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Drone'
  '/drone/{droneId}':
    get:
      tags:
        - Drone
      description: Returns drone by id
      operationId: getDrone
      parameters:
        - name: droneId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Get drone by id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Drone'
        '400':
          description: Bad requested
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataError'
  '/drone/{droneIdToUpdate}':
    put:
      tags:
        - Drone
      description: Update drone by id
      operationId: updateDrone
      parameters:
        - name: droneIdToUpdate
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DroneDto'
      responses:
        '200':
          description: Updated drone
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Drone'
        '400':
          description: Bad requested
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataError'
  '/drone/{droneIdToDelete}':
    delete:
      tags:
        - Drone
      description: Delete drone by id
      operationId: deleteDrone
      parameters:
        - name: droneIdToDelete
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: Drone with provided id
        '400':
          description: Bad requested
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataError'
...
components:
  schemas:
    Drone:
      type: object
      properties:
        id:
          format: int32
          type: integer
        model:
          type: string
        lastComponentsReview:
          format: date
          type: string
        latitude:
          type: number
        longitude:
          type: number
        deliveries:
          type: array
          items:
            $ref: '#/components/schemas/Delivery'
      description: ok
    DroneDto:
      type: object
      required: [model, latitude, longitude]
      properties:
        model:
          type: string
        latitude:
          type: number
        longitude:
          type: number
      description: ok
    Delivery:
      type: object
      properties:
        id:
          format: int32
          type: integer
        deliveryType:
          type: string
        deliveryStatus:
          type: string
        createOrder:
          format: date
          type: string
        lastUpdate:
          format: date
          type: string
        deliveredConfirmationLink:
          type: string
        drone:
          $ref: '#/components/schemas/Drone'
      description: ok
...
    DataError:
      type: object
      properties:
        error:
          type: string